@@ -53,21 +53,21 @@
   },
   "private": true,
   "dependencies": {
-    "@angular-redux/router": "^9.0.0",
-    "@angular-redux/store": "^9.0.0",
-    "@angular/animations": "^7.2.15",
+    "@angular-redux/router": "^10.0.0",
+    "@angular-redux/store": "^10.0.0",
+    "@angular/animations": "^8.0.0",
     "@angular/cdk": "^7.3.7",
-    "@angular/common": "^7.2.15",
-    "@angular/compiler": "^7.2.15",
-    "@angular/core": "^7.2.15",
-    "@angular/forms": "^7.2.15",
+    "@angular/common": "^8.0.0",
+    "@angular/compiler": "^8.0.0",
+    "@angular/core": "^8.0.0",
+    "@angular/forms": "^8.0.0",
     "@angular/material": "^7.3.7",
     "@angular/material-moment-adapter": "^7.3.7",
-    "@angular/platform-browser": "^7.2.15",
-    "@angular/platform-browser-dynamic": "^7.2.15",
-    "@angular/router": "^7.2.15",
+    "@angular/platform-browser": "^8.0.0",
+    "@angular/platform-browser-dynamic": "^8.0.0",
+    "@angular/router": "^8.0.0",
     "@angular/pwa": "^0.13.9",
-    "@angular/service-worker": "^7.2.15",
+    "@angular/service-worker": "^8.0.0",
     "@ngx-translate/core": "^11.0.1",
     "@redux-offline/redux-offline": "^2.5.1",
     "angular-calendar": "^0.27.8",
@@ -82,19 +82,19 @@
     "redux": "^4.0.1",
     "redux-observable": "^1.1.0",
     "rxjs": "^6.5.2",
-    "zone.js": "^0.9.1"
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.13.9",
-    "@angular-devkit/build-ng-packagr": "~0.13.9",
-    "@angular/cli": "7.3.9",
-    "@angular/compiler-cli": "~7.2.15",
-    "@angular/language-service": "^7.2.15",
+    "@angular-devkit/build-angular": "~0.800.0",
+    "@angular-devkit/build-ng-packagr": "~0.800.0",
+    "@angular/cli": "8.0.0",
+    "@angular/compiler-cli": "~8.0.0",
+    "@angular/language-service": "^8.0.0",
     "@nrwl/builders": "7.8.6",
     "@nrwl/schematics": "7.8.6",
     "@types/jest": "24.0.13",
     "@types/node": "~12.0.2",
-    "codelyzer": "~5.0.1",
+    "codelyzer": "^5.0.1",
     "cypress": "3.3.1",
     "dotenv": "8.0.0",
     "gulp-merge-json": "^1.3.1",
@@ -109,7 +109,7 @@
     "tsickle": ">=0.35.0",
     "tslib": "^1.9.3",
     "tslint": "~5.16.0",
-    "typescript": "3.2.4",
+    "typescript": "3.4.5",
     "webpack-bundle-analyzer": "^3.3.2"
   }
 }
